Move qcodes.dataset.database code to sqlite.database
qcodes.dataset.database now contains only import statements for
backwards compatibility. in following commits, all QCoDeS will use
sqlite.database instead of qcodes.dataset.database.
